// LDC.Atlas.WebApp/Client/app/shared/services/execution/dtos/contracts-to-washout-invoice.ts
export class ContractsToWashoutInvoice {
    sectionId: number;
    dataVersionId: number;
    contractReference: string;
    contractType: number;
    counterparty: string;
    principalCommodity: string;
    contractTermCode: string;
    allocatedContract: string;
    charterReference: string;
    quantityCode: string;
    quantity: number;
    currencyCode: string;
    priceCode: string;
    price: number;
    paymentTermCode: string;
    invoicedQuantity: number;
    blDate: Date;
    contractDate: Date;
    totalInvoiceValue: number;
    commodity: string;
    commodity2: string;
    commodity3: string;
    commodity4: string;
    commodity5: string;
    department: string;
    premiumDiscountType: number;
    premiumDiscountBasis: number;
    premiumDiscountValue: number;
    weightConversionFactor: number;
    priceConversionFactor: number;
}